uwatch: kill uwatch gracefully
There is a case when the uwatch is not a parent process, e.g. we have a
make file with a target:

```
    start:
            `$(GOPATH)`/bin/uwatch {some_app}
```

Then you run: `$ make start` and then you'll try to kill the process,
but uwatch changes the pgid and it won't get the death signal. To handle
this case we imitate a fork mechanism using an env variable to
distinguish parent and child processes.